Fix potential deadlock in worker->coordinator channels
